Since PostgreSQL doesn't have unsigned integers, use signed bigints to store u32 as i64

+++ res/schema.sql
@@ -1,23 +1,27 @@
+-- PostgreSQL doesn't have unsigned integers, therefore we have to use BIGINT (Signed 8 bytes, i64)
+-- to store u32 (Unsigned 4 bytes). Wasteful, i know.
+-- TODO: Find a way to convert u32 to i32 using traits or smth, they should be equivalent.
+
 CREATE TABLE accounts (
-    member_no INTEGER PRIMARY KEY,
+    member_no BIGINT PRIMARY KEY,
     login_id VARCHAR(20) NOT NULL UNIQUE,
     auth_key VARCHAR(64)
 );
 
 CREATE TABLE character (
-    member_no INTEGER ,
-    character_id INTEGER PRIMARY KEY,
+    member_no BIGINT ,
+    character_id BIGINT PRIMARY KEY,
     nickname VARCHAR(64) NOT NULL UNIQUE, -- TODO: Check length limit
-    char_id INTEGER NOT NULL,
-    mouth_serial_id INTEGER NOT NULL,
-    face_serial_id INTEGER NOT NULL,
-    parts_val0 INTEGER NOT NULL,
-    appearance_val0 INTEGER NOT NULL,
-    head_size INTEGER NOT NULL,
-    height INTEGER NOT NULL,
-    thigh_volume INTEGER NOT NULL,
-    leg_volume INTEGER NOT NULL,
-    val1 INTEGER NOT NULL,
-    create_character_unk0 INTEGER NOT NULL,
+    char_id BIGINT NOT NULL,
+    mouth_serial_id BIGINT NOT NULL,
+    face_serial_id BIGINT NOT NULL,
+    parts_val0 BIGINT NOT NULL,
+    appearance_val0 BIGINT NOT NULL,
+    head_size BIGINT NOT NULL,
+    height BIGINT NOT NULL,
+    thigh_volume BIGINT NOT NULL,
+    leg_volume BIGINT NOT NULL,
+    val1 BIGINT NOT NULL,
+    create_character_unk0 BIGINT NOT NULL,
     CONSTRAINT fk_member_no FOREIGN KEY (member_no) REFERENCES accounts(member_no)
 );
